+++ src/verify_cert.rs
@@ -408,101 +408,3 @@ where
     }
     Err(Error::UnknownIssuer.into())
 }
-
-#[cfg(feature = "alloc")]
-#[cfg(test)]
-mod tests {
-    use alloc::string::ToString;
-    use alloc::vec::Vec;
-    use core::convert::TryFrom;
-
-    use super::*;
-
-    enum ChainTrustAnchor {
-        InChain,
-        NotInChain,
-    }
-
-    fn build_degenerate_chain(
-        intermediate_count: usize,
-        trust_anchor: ChainTrustAnchor,
-    ) -> ErrorExt {
-        use crate::ECDSA_P256_SHA256;
-        use crate::{EndEntityCert, Time};
-
-        let alg = &rcgen::PKCS_ECDSA_P256_SHA256;
-
-        let make_issuer = |org_name| {
-            let mut ca_params = rcgen::CertificateParams::new(Vec::new());
-            ca_params
-                .distinguished_name
-                .push(rcgen::DnType::OrganizationName, org_name);
-            ca_params.is_ca = rcgen::IsCa::Ca(rcgen::BasicConstraints::Unconstrained);
-            ca_params.key_usages = vec![
-                rcgen::KeyUsagePurpose::KeyCertSign,
-                rcgen::KeyUsagePurpose::DigitalSignature,
-                rcgen::KeyUsagePurpose::CrlSign,
-            ];
-            ca_params.alg = alg;
-            rcgen::Certificate::from_params(ca_params).unwrap()
-        };
-
-        let ca_cert = make_issuer("Bogus Subject");
-        let ca_cert_der = ca_cert.serialize_der().unwrap();
-
-        let mut intermediates = Vec::with_capacity(intermediate_count);
-        if let ChainTrustAnchor::InChain = trust_anchor {
-            intermediates.push(ca_cert_der.to_vec());
-        }
-
-        let mut issuer = ca_cert;
-        for _ in 0..intermediate_count {
-            let intermediate = make_issuer("Bogus Subject");
-            let intermediate_der = intermediate.serialize_der_with_signer(&issuer).unwrap();
-            intermediates.push(intermediate_der);
-            issuer = intermediate;
-        }
-
-        let mut ee_params = rcgen::CertificateParams::new(vec!["example.com".to_string()]);
-        ee_params.is_ca = rcgen::IsCa::ExplicitNoCa;
-        ee_params.alg = alg;
-        let ee_cert = rcgen::Certificate::from_params(ee_params).unwrap();
-        let ee_cert_der = ee_cert.serialize_der_with_signer(&issuer).unwrap();
-
-        let trust_anchor = match trust_anchor {
-            ChainTrustAnchor::InChain => make_issuer("Bogus Trust Anchor").serialize_der().unwrap(),
-            ChainTrustAnchor::NotInChain => ca_cert_der.clone(),
-        };
-
-        let anchors = &[TrustAnchor::try_from_cert_der(&trust_anchor).unwrap()];
-        let time = Time::from_seconds_since_unix_epoch(0x1fed_f00d);
-        let cert = EndEntityCert::try_from(&ee_cert_der[..]).unwrap();
-        let intermediate_certs = intermediates.iter().map(|x| x.as_ref()).collect::<Vec<_>>();
-
-        build_chain_inner(
-            EKU_SERVER_AUTH,
-            &[&ECDSA_P256_SHA256],
-            anchors,
-            &intermediate_certs,
-            cert.inner(),
-            time,
-            0,
-            &mut Budget::default(),
-        )
-        .unwrap_err()
-    }
-
-    #[test]
-    fn test_too_many_signatures() {
-        assert!(matches!(
-            build_degenerate_chain(5, ChainTrustAnchor::NotInChain),
-            ErrorExt::MaximumSignatureChecksExceeded,
-        ));
-    }
-
-    #[test]
-    fn test_too_many_path_calls() {
-        let result = build_degenerate_chain(10, ChainTrustAnchor::InChain);
-        assert!(matches!(result, ErrorExt::MaximumPathBuildCallsExceeded,));
-    }
-}
